Move Margins widget to reserve module
